# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Under the Radar API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "username", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "email": { "type": "string" },
                  "bio": { "type": "string" },
                  "social": { "type": "string" },
                  "is_artist": { "type": "boolean" }
                }
              },
              "example": {
                "name": "Alice",
                "username": "alice",
                "password": "secret",
                "is_artist": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token",
            "content": {
              "application/json": {
                "example": { "token": "jwt", "id": 1 }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and obtain a token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              },
              "example": {
                "username": "alice",
                "password": "secret"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token",
            "content": {
              "application/json": {
                "example": { "token": "jwt", "id": 1 }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "List users",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string", "enum": ["artist", "user"] },
            "description": "Filter by profile type"
          },
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search name or username"
          },
          {
            "name": "letter",
            "in": "query",
            "schema": { "type": "string", "minLength": 1, "maxLength": 1 },
            "description": "Filter by starting letter"
          }
        ],
        "responses": {
          "200": {
            "description": "User list",
            "content": {
              "application/json": {
                "example": [
                  {"id": 1, "name": "Alice"}
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update current user profile",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"type": "object"}}}
        },
        "responses": {"200": {"description": "Update info"}}
      }
    },
    "/users/avatar": {
      "post": {
        "summary": "Upload avatar image",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {"type": "string", "format": "binary"}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "Avatar info"}}
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Fetch a user by id",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "User", "content": {"application/json": {"example": {"id": 1, "name": "Alice"}}}}}
      }
    },
    "/messages/inbox": {
      "get": {
        "summary": "Get inbox messages",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "Messages"}}
      }
    },
    "/messages/outbox": {
      "get": {
        "summary": "Get outbox messages",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "Messages"}}
      }
    },
    "/messages": {
      "post": {
        "summary": "Send a message",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["receiver_id", "content"],
                "properties": {
                  "receiver_id": { "type": "integer" },
                  "content": { "type": "string" }
                }
              },
              "example": {"receiver_id": 2, "content": "Hi"}
            }
          }
        },
        "responses": {"200": {"description": "Message created"}}
      }
    },
    "/media": {
      "get": {
        "summary": "List uploaded files",
        "responses": {"200": {"description": "Files"}}
      },
      "post": {
        "summary": "Upload a file",
        "security": [{"BearerAuth": []}],
        "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"type": "object", "properties": {"file": {"type": "string", "format": "binary"}}}}}},
        "responses": {"200": {"description": "File metadata"}}
      }
    },
    "/media/{id}": {
      "get": {
        "summary": "Download or stream a file",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "File"}}
      }
    },
    "/profile-media": {
      "post": {
        "summary": "Upload profile picture or video",
        "security": [{"BearerAuth": []}],
        "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"type": "object", "properties": {"file": {"type": "string", "format": "binary"}}}}}},
        "responses": {"200": {"description": "Created"}}
      }
    },
    "/profile-media/user/{id}": {
      "get": {
        "summary": "Profile media for user",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Files"}}
      }
    },
    "/shows": {
      "get": {"summary": "List shows", "responses": {"200": {"description": "Shows"}}},
      "post": {"summary": "Create a show", "security": [{"BearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["venue", "date"], "properties": {"venue": {"type": "string"}, "date": {"type": "string"}, "description": {"type": "string"}}}}}}, "responses": {"200": {"description": "Created"}}}
    },
    "/shows/user/{id}": {
      "get": {"summary": "Shows for a user", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Shows"}}}
    },
    "/merch": {
      "get": {"summary": "List merch", "responses": {"200": {"description": "Items"}}},
      "post": {"summary": "Create merch", "security": [{"BearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["product_name", "price"], "properties": {"product_name": {"type": "string"}, "price": {"type": "number"}, "stock": {"type": "integer"}}}}}}, "responses": {"200": {"description": "Created"}}}
    },
    "/merch/user/{id}": {
      "get": {"summary": "Merch for user", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Items"}}}
    },
    "/board": {
      "get": {"summary": "List board posts", "description": "Posts include `updated_at` when edited", "responses": {"200": {"description": "Posts"}}},
      "post": {"summary": "Create board post", "security": [{"BearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["headline", "content"], "properties": {"headline": {"type": "string"}, "content": {"type": "string"}}}}}}, "responses": {"200": {"description": "Created"}}}
    },
    "/board/user/{id}": {
      "get": {"summary": "Posts by user", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Posts"}}}
    },
    "/board/{id}/like": {
      "post": {"summary": "Like a post", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Success"}}}
    },
    "/board/{id}/dislike": {
      "post": {"summary": "Dislike a post", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Success"}}}
    },
    "/board/{id}/comments": {
      "get": {"summary": "List comments", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Comments"}}},
      "post": {"summary": "Add comment", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["content"], "properties": {"content": {"type": "string"}}}}}}, "responses": {"200": {"description": "Created"}}}
    },
    "/board/comments/{id}": {
      "put": {"summary": "Edit comment", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["content"], "properties": {"content": {"type": "string"}}}}}}, "responses": {"200": {"description": "Updated"}}},
      "delete": {"summary": "Delete comment", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Deleted"}}}
    },
    "/board/{id}": {
      "put": {"summary": "Edit post", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["headline", "content"], "properties": {"headline": {"type": "string"}, "content": {"type": "string"}}}}}}, "responses": {"200": {"description": "Updated"}}},
      "delete": {"summary": "Delete post", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Deleted"}}}
    },
    "/follow/followers/{id}": {
      "get": {"summary": "List followers", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Followers"}}}
    },
    "/follow/following/{id}": {
      "get": {"summary": "List following", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Following"}}}
    },
    "/follow/{id}": {
      "get": {"summary": "Check follow", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Status"}}},
      "post": {"summary": "Follow user", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Followed"}}},
      "delete": {"summary": "Unfollow user", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Unfollowed"}}}
    },
    "/notifications": {
      "get": {"summary": "List notifications", "security": [{"BearerAuth": []}], "responses": {"200": {"description": "Notifications"}}}
    },
    "/notifications/{id}/read": {
      "post": {"summary": "Mark notification read", "security": [{"BearerAuth": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Updated"}}}
    },
    "/health": {
      "get": {"summary": "Health check", "responses": {"200": {"description": "OK", "content": {"application/json": {"example": {"status": "ok"}}}}}}
    },
    "/metrics": {
      "get": {"summary": "Basic metrics", "responses": {"200": {"description": "Metrics"}}}
    }
  }
}